@@ -3,33 +3,44 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@mui/material": "^5.0.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.0",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/jest": "^27.0.1",
-    "@types/node": "^16.9.2",
-    "@types/react": "^17.0.21",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.1.9",
-    "@types/styled-components": "^5.1.14",
-    "@types/webpack-env": "^1.16.2",
-    "material-ui": "^0.20.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-photo-gallery": "^8.0.0",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.3.1",
-    "typescript": "~4.4.3"
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@mui/material": "^6.3.0",
+    "@testing-library/jest-dom": "latest",
+    "@testing-library/react": "latest",
+    "@testing-library/user-event": "latest",
+    "@types/jest": "latest",
+    "@types/node": "latest",
+    "@types/react": "latest",
+    "@types/react-dom": "latest",
+    "@types/react-router-dom": "latest",
+    "@types/styled-components": "latest",
+    "@types/webpack-env": "latest",
+    "axios": "latest",
+    "buffer": "^6.0.3",
+    "dotenv": "latest",
+    "file-loader": "latest",
+    "google-auth-library": "latest",
+    "googleapis": "latest",
+    "http2": "latest",
+    "process": "^0.11.10",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-responsive": "latest",
+    "react-router-dom": "latest",
+    "react-scripts": "^5.0.1",
+    "react-slideshow-image": "latest",
+    "stream-browserify": "^3.0.0",
+    "styled-components": "latest",
+    "typescript": "latest",
+    "webpack": "latest",
+    "webpack-cli": "latest"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test",
+    "eject": "react-app-rewired eject"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -46,18 +57,11 @@
       "last 1 safari version"
     ]
   },
-  "description": "All of these commands are run in terminal!",
-  "main": "index.js",
-  "devDependencies": {},
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/vsubruins/vsu-website.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/vsubruins/vsu-website/issues"
-  },
-  "homepage": "https://github.com/vsubruins/vsu-website#readme"
+  "devDependencies": {
+    "crypto-browserify": "^3.12.1",
+    "dotenv-webpack": "^8.1.0",
+    "os-browserify": "^0.3.0",
+    "path-browserify": "^1.0.1",
+    "react-app-rewired": "^2.2.1"
+  }
 }
