@@ -3,8 +3,8 @@ import styled from "styled-components";
 
 import SocialMediaIcons from "../SocialMediaIcons";
 
-import { ReactComponent as YellowCircle } from "../../assets/yellowcircle.svg";
-import { ReactComponent as BlueTriangle } from "../../assets/bluetriangle.svg";
+import { ReactComponent as YellowCircle } from "../../assets/shapes/yellowcircle.svg";
+import { ReactComponent as BlueTriangle } from "../../assets/shapes/bluetriangle.svg";
 
 const Wrapper = styled("div")`
   display: flex;
@@ -15,13 +15,14 @@ const Wrapper = styled("div")`
 `;
 
 const FollowUs = styled("p")`
-  font-size: 30px;
+  font-size: 2rem;
   letter-spacing: 0.05em;
+  text-align: center;
 `;
 
 const Circle = styled(YellowCircle)`
   position: absolute;
-  left: -90px;
+  left: -2rem;
   top: -40px;
   width: 110px;
   z-index: -100;
@@ -31,7 +32,7 @@ const Triangle = styled(BlueTriangle)`
   position: absolute;
   z-index: -100;
   width: 120px;
-  right: -100px;
+  right: -3vw;
 `;
 
 export default function LandingSocialMedia() {
