@@ -13,18 +13,24 @@ const NavWrapper = styled("div")`
   height: 76.32px;
   margin-top: 30px;
   margin-bottom: 30px;
+
+  @media (max-width: 900px) {
+    margin-bottom: 50px;
+  }
 `;
+
 
 const LogoDiv = styled(Link)`
   position: absolute;
   left: 50%;
-  /* margin-top: 30px;
-  margin-bottom: 30px; */
 `;
 
 const VSULogo = styled(Logo)`
   position: relative;
   left: -50%;
+  height: 6rem;
+  width: 6rem;
+  margin-top: -10px;
 `;
 
 export default function NavBar() {
