@@ -1,58 +1,100 @@
 import React from "react";
 import styled from "styled-components";
-
 import { Link } from "react-router-dom";
+import DropdownMenu from "./DropdownMenu"; // Imported DropdownMenu
 
 const Wrapper = styled("div")`
   display: flex;
   align-items: center;
-  margin-left: 10px;
+  margin-left: 30px;
+
+  @media (max-width: 900px) {
+    margin-top: 100px;
+  }
 `;
 
 const NavLink = styled(Link)`
   text-decoration: none;
   color: black;
-  padding: 15px 8px;
+  padding: 0px 8px;
   margin-left: 5px;
   margin-right: 5px;
-  font-size: 14px;
-
-  &:hover {
-    border-bottom: black 2px solid;
-    padding-bottom: 13px;
-  }
+  font-size: 16px;
 `;
 
-const ExtNavLink = styled("a")`
-  text-decoration: none;
-  color: black;
-  padding: 15px 8px;
-  margin-left: 5px;
-  margin-right: 5px;
-  font-size: 14px;
+type LinkItem = {
+  to: string;
+  name: string;
+};
 
-  &:hover {
-    border-bottom: black 2px solid;
-    padding-bottom: 13px;
-  }
-`;
+type DropdownMenuProps = {
+  header: LinkItem;
+  list: LinkItem[];
+};
+
+const DropdownMenuComponent = ({ header, list }: DropdownMenuProps) => {
+  return (
+    <div>
+      <Link to={header.to}>{header.name}</Link>
+      <ul>
+        {list.map((item, index) => (
+          <li key={index}>
+            <Link to={item.to}>{item.name}</Link>
+          </li>
+        ))}
+      </ul>
+    </div>
+  );
+};
 
 export default function Navigation() {
+  const aboutHeader: LinkItem = {
+    to: "/about",
+    name: "About",
+  };
+
+  const aboutList: LinkItem[] = [
+    // { to: "/staff", name: "Staff" }
+  ];
+
+  const involvementHeader: LinkItem = {
+    to: "/membership",
+    name: "Membership",
+  };
+
+  const involvementList: LinkItem[] = [
+    { to: "/internship", name: "Internship" },
+    { to: "/seaclear", name: "SEA CLEAR" },
+    { to: "/hope", name: "HOPE" },
+  ];
+
+  const programsHeader: LinkItem = {
+    to: "/",
+    name: "Our Programs",
+  };
+
+  const programsList: LinkItem[] = [
+    { to: "/culturenight", name: "Culture Night" },
+    { to: "/educationprograms", name: "Education Programs" },
+  ];
+
+  const famsHeader: LinkItem = {
+    to: "/JoinAFam",
+    name: "Fams",
+  };
+
+  const famsList: LinkItem[] = [
+    { to: "/MeetTheBigs", name: "Meet The Bigs!" },
+    { to: "/JoinAFam", name: "Join A Fam!" },
+  ];
+
   return (
     <>
       <Wrapper>
-        <NavLink to="/about">About</NavLink>
-        <ExtNavLink
-          target="_blank"
-          href="https://www.facebook.com/vsubruins"
-        >
-          Events
-        </ExtNavLink>
-        <NavLink to="/hope">HOPE</NavLink>
-        <NavLink to="/seaclear">SEA CLEAR</NavLink>
-        <NavLink to="/staff">Staff</NavLink>
-        <NavLink to="/gallery">Gallery</NavLink>
-        <NavLink to="/membership">Membership</NavLink>
+        <DropdownMenu header={aboutHeader} list={aboutList} />
+        <DropdownMenu header={involvementHeader} list={involvementList} />
+        <DropdownMenu header={programsHeader} list={programsList} />
+        <DropdownMenu header={famsHeader} list={famsList} />
       </Wrapper>
     </>
   );
