@@ -9,6 +9,21 @@ import {
   PageWrapper,
   PageContent,
 } from "../components/Page/Page";
+import styled from "styled-components";
+
+const ExtNavLink = styled("a")`
+  text-decoration: none;
+  color: black;
+  padding: 15px 8px;
+  margin-left: 5px;
+  margin-right: 5px;
+  font-size: 18px;
+
+  &:hover {
+    border-bottom: black 2px solid;
+    padding-bottom: 13px;
+  }
+`;
 
 export default function HOPE() {
   return (
@@ -19,30 +34,44 @@ export default function HOPE() {
           <PageTitle>Higher Opportunity Program for Education</PageTitle>
           <PageText>
             The Higher Opportunity Program for Education (HOPE) is a
-            student-initiated project founded by the Vietnamese Student Union
-            (VSU) in 1999. HOPE is committed to outreaching to under-resourced
-            and underrepresented youth. HOPE exists to raise consciousness of
-            and provide access to higher education through holistic empowerment.
-            We promote self-sufficiency and leadership through academic,
-            personal, and community empowerment. We develop our students through
-            peer-advising, tutoring, workshops, college tours, and field trips.
-            Under the support of the Student-Initiated Outreach Committee (SIOC)
-            and the direction of VSU, HOPE strives to achieve its goals toward
-            educational equity for all.
+            student-initiated project founded by UCLA's Vietnamese Student Union
+            (VSU) in 1999. youths who face educational obstacles and exists to raise 
+            consciousness of and provide access to higher education through holistic 
+            empowerment. With the support of the Student Initiated Outreach Committee 
+            (SIOC) and the direction of VSU, HOPE strives to achieve its goals toward 
+            educational equity for all by providing students with academic tutoring 
+            services, mentoring and peer advising sessions, workshops on various topics, 
+            & educational field trips.
           </PageText>
           <PageHeader>When does HOPE meet?</PageHeader>
           <PageText>
             <ul>
-              <li>HOPE All Staff/Volunteer Meetings, Mondays 6-8pm</li>
-              <li>Westminster High School Site, Tuesdays 2:00-7:00pm</li>
-              <li>San Gabriel High School Site, Wednesdays 1:00-7:00pm</li>
+              <li>Westminster High School Site, Tuesdays 2:00 PM - 8:00 PM</li>
+              <li>Gilbert High School Site, Wednesdays 12:00 PM - 6:00 PM</li>
+              <li>Loara High School Site, Thursdays 2:00 PM - 8:00 PM</li>
+              <li>Field trips:
+                <ul>
+                  <li>Shadow Days - HOPE hosts one Shadow Day for each partnered school, letting 
+                      students explore UCLA’s campus and dorms with real college mentors</li>
+                  <li>College Tour - HOPE takes their high school students to tour different 
+                      college campuses and learn the different types of higher education</li>
+                  <li>Community Tour - HOPE gives students the tools to celebrate their identity, 
+                      promote self-care, and support themselves holistically</li>
+                  <li>Banquet - An end of the year celebration for HOPE volunteers and students to 
+                      recognize the year’s accomplishments</li>
+                </ul>
+              </li>
             </ul>
             Note: Site times include transportation time
           </PageText>
           <PageHeader>Services</PageHeader>
           <PageText>
             <ul>
-              <li>Peer Advising</li>
+              <li>Peer Advising
+                <ul>
+                  <li>Mentorship</li>
+                </ul>
+              </li>
               <li>
                 Academic Development{" "}
                 <ul>
@@ -52,22 +81,24 @@ export default function HOPE() {
                   <li>Skill-building Workshops</li>
                 </ul>
               </li>
-              <li>
-                Collective Development
+              <li>Higher Education
                 <ul>
-                  <li>
-                    Collectively Reaching Empowerment through Workshops (CREW)
-                  </li>
-                  <li>Reaching Independence through Self Efficacy (RISE)</li>
+                  <li>Different routes of higher education</li>
                 </ul>
               </li>
-              <li>Parent Engagement</li>
+              <li>Internship
+                <ul>
+                  <li>Professional development</li>
+                  <li>Retreats</li>
+                  <li>Community</li>
+                </ul>
+              </li>
             </ul>
           </PageText>
           <PageHeader>Peer Advising</PageHeader>
           <PageText>
-            Through peer-mentoring, HOPE provides an outlet for students to
-            express themselves and help students through high school and plan
+            Through peer advising, HOPE provides an outlet for our high school students to
+            express themselves and help them get through high school and plan
             for life after graduating. To prepare them for college, sessions get
             students to understand A-G requirements, keep them accountable for
             high school graduation, and give students information on tests
@@ -87,8 +118,8 @@ export default function HOPE() {
           <PageHeader>Academic Development</PageHeader>
           <PageText>
             The Academic Development component is not just about homework help.
-            It is about developing students academically so that they can be
-            successful in whatever they do. <br />
+            It is about developing our high school students academically so that 
+            they can be successful in whatever they do. <br />
             <br />
             Students will grow academically through tutoring, skill building,
             test prep, and writing development. Tutoring is one of the biggest
@@ -97,25 +128,9 @@ export default function HOPE() {
             student develop their understanding of the process and the writing
             structure. This can be done via journals, blogs, etc.
           </PageText>
-          <PageHeader>Collective Development</PageHeader>
-          <PageText>
-            <u>CREW</u>: Reaching Empowerment through Workshops <br />
-            <br />
-            CREW workshops touch on topics that can only be learned outside the
-            classroom. Here, students are exposed to community issues, personal
-            wellness, current events, politics, as well as topics they choose.{" "}
-            <br />
-            <br />
-            <u>RISE</u>: Reaching Independence through Self-Efficacy <br />
-            <br />
-            RISE component exposes students to college and help them to
-            effectively choose, apply, and prepare for higher education. RISE
-            includes college tours, shadow days, and college preparation
-            workshops.
-          </PageText>
           <PageHeader>Internship</PageHeader>
           <PageText>
-            Internship provides an opportunity for students to learn more about
+            Internship provides an opportunity for UCLA students to learn more about
             themselves, their communities, and actions for change. The HOPE
             internship will enable interns to critically think about issues that
             affect access to higher education as well as their identity and
@@ -124,16 +139,26 @@ export default function HOPE() {
             professional growth, allowing them to be more prepared to take on
             leadership roles within and outside the HOPE space.
           </PageText>
+          {/* UNCOMMENT + UPDATE FOR WELCOME WEEK
+          <ExtNavLink
+            target="_blank"
+            href="http://tinyurl.com/hope-internship"
+          >
+            HOPE Internship Application
+          </ExtNavLink>
+          */}
           <PageHeader>Contact</PageHeader>
+          <br />
+          <ExtNavLink
+            target="_blank"
+            href="http://tinyurl.com/HOPE-Interest"
+            >
+            Volunteer Interest Form
+          </ExtNavLink>
           <PageText>
-            <ul>
-              <li>Call: 310-267-4346</li>
-              <li>Email: hope@cpo.ucla.edu</li>
-              <li>
-                Our office is located in the Student Activities Center (SAC), in
-                Room 106D
-              </li>
-            </ul>
+            Our office is located in the Student Activities Center (SAC), in Room 106D.<br />
+            <b>Phone:</b> 310-267-4346<br />
+            <b>Email:</b> hope@cpo.ucla.edu
           </PageText>
         </PageContent>
         <Footer />
